salvar alterado

+++ src/main/java/application/service/JogoService.java
@@ -1,7 +1,11 @@
 package application.service;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import application.model.Jogo;
 import application.record.JogoDTO;
@@ -12,6 +16,7 @@ import application.repository.JogoRepository;
 public class JogoService {
     @Autowired
     private JogoRepository jogoRepo;
+
 
     public Iterable<JogoListDTO> getAllSimple() {
         return jogoRepo.findAll().stream().map(JogoListDTO::new).toList();
@@ -25,7 +30,24 @@ public class JogoService {
         return new JogoDTO(jogoRepo.save(new Jogo(jogo)));
     }
 
+    public JogoDTO update(long id, JogoDTO jogo) {
+        Optional<Jogo> result = jogoRepo.findById(id);
+        if(result.isEmpty()) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Jogo"
+            );
+        }
+        result.get().setPlataformas(jogo.plataformas());
+        result.get().setTitulo(jogo.titulo());
+ 
+        return new JogoDTO(jogoRepo.save(result.get()));
+    }
+
     public void deleteById(long id) {
+        if (jogoRepo.existsById(id)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Jogo Não Encontrado");
+        }
         jogoRepo.deleteById(id);
     }
 }
